Move EVMHandler
